@@ -23,8 +23,8 @@ import com.erudika.scoold.utils.ScooldUtils;
 import java.io.IOException;
 import java.util.Collections;
 import javax.inject.Inject;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
@@ -50,14 +50,14 @@ public class ErrorController {
 	public String get(@PathVariable String code, HttpServletRequest req, HttpServletResponse res, Model model) throws IOException {
 		model.addAttribute("path", "error.vm");
 		model.addAttribute("title", utils.getLang(req).get("error.title"));
-		model.addAttribute("status", req.getAttribute("javax.servlet.error.status_code"));
-		model.addAttribute("reason", req.getAttribute("javax.servlet.error.message"));
+		model.addAttribute("status", req.getAttribute("jakarta.servlet.error.status_code"));
+		model.addAttribute("reason", req.getAttribute("jakarta.servlet.error.message"));
 		model.addAttribute("code", code);
 
-		if (StringUtils.startsWith((CharSequence) req.getAttribute("javax.servlet.forward.request_uri"), "/api/")) {
+		if (StringUtils.startsWith((CharSequence) req.getAttribute("jakarta.servlet.forward.request_uri"), "/api/")) {
 			res.setContentType(MediaType.APPLICATION_JSON_VALUE);
 			ParaObjectUtils.getJsonWriterNoIdent().writeValue(res.getOutputStream(),
-					Collections.singletonMap("error", code + " - " + req.getAttribute("javax.servlet.error.message")));
+					Collections.singletonMap("error", code + " - " + req.getAttribute("jakarta.servlet.error.message")));
 		}
 		return "base";
 	}
