@@ -19,7 +19,7 @@ package com.erudika.scoold.controllers;
 
 import com.erudika.scoold.utils.ScooldUtils;
 import javax.inject.Inject;
-import javax.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
