@@ -23,8 +23,8 @@ import com.erudika.scoold.ScooldConfig;
 import java.io.InputStream;
 import java.util.Iterator;
 import java.util.List;
-import javax.mail.internet.MimeMessage;
-import javax.mail.util.ByteArrayDataSource;
+import jakarta.mail.internet.MimeMessage;
+import jakarta.mail.util.ByteArrayDataSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.mail.MailException;
